requirements.txt
Django~=3.1.3
Pillow~=8.0.1
pip~=20.2.4
setuptools~=50.3.2
django-crispy-forms==1.10.0
pytz==2020.4
sqlparse==0.4.1
djangorestframework==3.12.2